requirements.txt
langchain
langchain-google-genai
langchain-community
langgraph
yfinance
gradio
requests
python-dotenv
